Use bun for testing
