--- ava.config.mjs
@@ -1,4 +1,0 @@
-export default {
-	extensions: ['ts'],
-	nodeArguments: ['--import=tsx'],
-};
