+++ eslint.config.mjs
@@ -1,7 +1,3 @@
-import path from 'node:path';
-import {fileURLToPath} from 'node:url';
-import {fixupConfigRules} from '@eslint/compat';
-import {FlatCompat} from '@eslint/eslintrc';
 import js from '@eslint/js';
 import tseslint from '@typescript-eslint/eslint-plugin';
 import tsParser from '@typescript-eslint/parser';
@@ -9,14 +5,6 @@ import eslintConfigPrettier from 'eslint-config-prettier';
 import eslintPluginUnicorn from 'eslint-plugin-unicorn';
 import {defineConfig} from 'eslint/config';
 import globals from 'globals';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const compat = new FlatCompat({
-	baseDirectory: __dirname,
-	recommendedConfig: js.configs.recommended,
-	allConfig: js.configs.all,
-});
 
 export default defineConfig([
 	{
@@ -27,8 +15,6 @@ export default defineConfig([
 	eslintConfigPrettier,
 
 	{
-		extends: fixupConfigRules(compat.extends('plugin:ava/recommended')),
-
 		files: ['**/*.ts', '**/*.tsx'],
 		languageOptions: {
 			parser: tsParser,
