+++ package.json
@@ -4,7 +4,7 @@
 	"description": "Check the source files in directories",
 	"scripts": {
 		"build": "tsc",
-		"test": "ava",
+		"test": "bun test",
 		"lint": "eslint .",
 		"format:check": "prettier . --check"
 	},
@@ -57,20 +57,16 @@
 		"@eslint/js": "^9.27.0",
 		"@ianvs/prettier-plugin-sort-imports": "^4.4.1",
 		"@tsconfig/node20": "^20.1.4",
+		"@types/bun": "^1.3.2",
 		"@types/node": "^22.10.5",
 		"@typescript-eslint/eslint-plugin": "^8.18.2",
 		"@typescript-eslint/parser": "^8.18.2",
-		"ava": "^6.3.0",
 		"eslint": "^9.27.0",
 		"eslint-config-prettier": "^10.1.5",
-		"eslint-plugin-ava": "^15.0.1",
 		"eslint-plugin-import": "^2.31.0",
 		"eslint-plugin-unicorn": "^59.0.1",
 		"globals": "^16.1.0",
-		"mock-fs": "^5.5.0",
 		"prettier": "^3.5.3",
-		"ts-node": "^10.9.2",
-		"tsx": "^4.19.2",
 		"typescript": "^5.7.2"
 	}
 }
