+++ tsconfig.json
@@ -14,8 +14,9 @@
 		"resolveJsonModule": true,
 		"declaration": true,
 		"declarationMap": true,
-		"sourceMap": true
+		"sourceMap": true,
+		"types": ["bun-types"]
 	},
 	"include": ["**/*.ts"],
-	"exclude": ["node_modules", "dist"]
+	"exclude": ["node_modules", "dist", "test"]
 }
